M - Program.cs - AppConfig inicialization moved higher corrected work of 4th menu option
M - JsonFIleService -
Added inicialization for file password.json
simplified access to AppConfig prop from file
logic issues corrected

M - UserDataManager -
Now user can see full table (with decrypted passwords) after enter an admin password
Now if app user want to find password by site name, at first he must eneter an admin password

M - AppConfig -
Added one more prop with admin password which is next encrypted
Also check is file exist for save some memory

+++ PasswordKeeper.App/Settings/AppConfig.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PasswordKeeper.Cypher;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http.Headers;
@@ -16,18 +17,26 @@ namespace PasswordKeeper.App.Settings
 
         public string SecurityKey { get; set; }
 
+        public string AdminPassword { get; set; }
+
         public void SetConfig()
         {
+            if (File.Exists("appsettings.json"))
+                return;
+
             var appSettings = new AppConfig
             {
                 PassordsFilePath = "passwords.json",
                 MaxAttempts = 3,
-                SecurityKey = "TMPSecK3yAmg1g18ait"
+                SecurityKey = "TMPSecK3yAmg1g18ait",
             };
-
             string jsonString = JsonSerializer.Serialize(appSettings);
+            File.WriteAllText("appsettings.json", jsonString);
 
+            appSettings.AdminPassword = EncryptionsMaker.EncryptPlainTextToCipherText("admin1");
+            jsonString = JsonSerializer.Serialize(appSettings);
             File.WriteAllText("appsettings.json", jsonString);
         }
     }
 }
+
